Updated config and patched muliti-config bug
