--- cli/config/cfg.go
@@ -1,5 +1,0 @@
-package config
-
-const (
-	CfgDir = ".vx"
-)
